+++ src/resources/database/migrations/2023_04_08_104809_create_languages_table.php
@@ -16,8 +16,6 @@ return new class () extends Migration {
                 $table->string('native_name');
 
                 $table->timestamps();
-
-                //$table->primary('id');
             });
         }
     }