+++ tests/AddressTest.php
@@ -201,9 +201,9 @@ class AddressTest extends TestCase
     /**
      * @inheritdoc
      */
-    protected function setUpDatabase($application)
+    protected function setUpDatabase($app)
     {
-        parent::setUpDatabase($application);
+        parent::setUpDatabase($app);
 
         $this->artisan('db:seed', ['--class' => Countries::class]);
         $this->artisan('db:seed', ['--class' => StatesOfUsa::class]);