+++ tests/ProvinceTest.php
@@ -151,9 +151,9 @@ class ProvinceTest extends TestCase
     /**
      * @inheritdoc
      */
-    protected function setUpDatabase($application)
+    protected function setUpDatabase($app)
     {
-        parent::setUpDatabase($application);
+        parent::setUpDatabase($app);
 
         $this->artisan('db:seed', ['--class' => Countries::class]);
         $this->artisan('db:seed', ['--class' => CountiesOfHungary::class]);