+++ tests/Seeders/BelgiumTest.php
@@ -123,9 +123,9 @@ class BelgiumTest extends TestCase
         $this->assertContains('Walloon Brabant', $names);
     }
 
-    protected function setUpDatabase($application)
+    protected function setUpDatabase($app)
     {
-        parent::setUpDatabase($application);
+        parent::setUpDatabase($app);
 
         $this->artisan('db:seed', ['--class' => Countries::class]);
         $this->artisan('db:seed', ['--class' => ProvincesAndRegionsOfBelgium::class]);