+++ tests/Seeders/HungaryTest.php
@@ -69,9 +69,9 @@ class HungaryTest extends TestCase
         $this->assertContains('Zala', $names);
     }
 
-    protected function setUpDatabase($application)
+    protected function setUpDatabase($app)
     {
-        parent::setUpDatabase($application);
+        parent::setUpDatabase($app);
 
         $this->artisan('db:seed', ['--class' => Countries::class]);
         $this->artisan('db:seed', ['--class' => CountiesOfHungary::class]);