+++ tests/Seeders/IndiaTest.php
@@ -99,9 +99,9 @@ class IndiaTest extends TestCase
         $this->assertContains('Puducherry', $names);
     }
 
-    protected function setUpDatabase($application)
+    protected function setUpDatabase($app)
     {
-        parent::setUpDatabase($application);
+        parent::setUpDatabase($app);
 
         $this->artisan('db:seed', ['--class' => Countries::class]);
         $this->artisan('db:seed', ['--class' => StatesAndTerritoriesOfIndia::class]);