+++ tests/Seeders/IndonesiaTest.php
@@ -209,9 +209,9 @@ class IndonesiaTest extends TestCase
         $this->assertContains('Riau Islands', $names);
     }
 
-    protected function setUpDatabase($application)
+    protected function setUpDatabase($app)
     {
-        parent::setUpDatabase($application);
+        parent::setUpDatabase($app);
 
         $this->artisan('db:seed', ['--class' => Countries::class]);
         $this->artisan('db:seed', ['--class' => ProvincesOfIndonesia::class]);