+++ tests/Seeders/NetherlandsTest.php
@@ -61,9 +61,9 @@ class NetherlandsTest extends TestCase
         $this->assertContains('Zuid-Holland', $names);
     }
 
-    protected function setUpDatabase($application)
+    protected function setUpDatabase($app)
     {
-        parent::setUpDatabase($application);
+        parent::setUpDatabase($app);
 
         $this->artisan('db:seed', ['--class' => Countries::class]);
         $this->artisan('db:seed', ['--class' => ProvincesOfNetherlands::class]);